+++ omnibenchmark/cli/describe.py
@@ -1,12 +1,9 @@
-"""cli commands related to benchmark status"""
+"""cli commands related to benchmark infos and stats"""
 
 import sys
 
 import click
 import json
-
-from omnibenchmark.cli.utils.logging import logger
-
 
 import pandas as pd
 from jinja2 import Environment, FileSystemLoader
@@ -16,12 +13,60 @@ from omnibenchmark.benchmark.status.status import prepare_status, print_exec_pat
 from omnibenchmark.benchmark import BenchmarkExecution
 from omnibenchmark.dag import get_node_attributes
 
-
+from omnibenchmark.cli.utils.logging import logger
 from .debug import add_debug_option
 
 
+@click.group(name="describe")
+@click.pass_context
+# @debug_option
+def describe(ctx):
+    """Describe benchmarks and/or information about them."""
+    ctx.ensure_object(dict)
+
+
 @add_debug_option
-@click.command(
+@describe.command("computational")
+@click.option(
+    "--benchmark",
+    "-b",
+    required=True,
+    type=click.Path(exists=True),
+    help="Path to benchmark yaml file or benchmark id.",
+    envvar="OB_BENCHMARK",
+)
+@click.pass_context
+def computational_graph(ctx, benchmark: str):
+    """Export computational graph to dot format."""
+    b = BenchmarkExecution(benchmark_yaml=Path(benchmark))
+    if b is None:
+        return
+    dot = b.export_to_dot()
+    click.echo(dot.to_string())
+
+
+@add_debug_option
+@describe.command("topology")
+@click.option(
+    "--benchmark",
+    "-b",
+    required=True,
+    type=click.Path(exists=True),
+    help="Path to benchmark yaml file or benchmark id.",
+    envvar="OB_BENCHMARK",
+)
+@click.pass_context
+def plot_topology(ctx, benchmark: str):
+    """Export benchmark topology to mermaid diagram format."""
+    b = BenchmarkExecution(benchmark_yaml=Path(benchmark))
+    if b is None:
+        return
+    mermaid = b.export_to_mermaid()
+    click.echo(mermaid)
+
+
+@add_debug_option
+@describe.command(
     name="status",
     context_settings=dict(
         ignore_unknown_options=False,